strict validation, fix config endpoints
